backend/requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
python-multipart==0.0.6
pydantic==1.10.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
selenium==4.8.0
beautifulsoup4==4.11.0
requests==2.28.0
pytest==7.2.0
httpx==0.24.0